src/scripts/data_preparation/alt/requirements_docker.txt
pandas
mysql-connector-python
beautifulsoup4
requests
selenium
webdriver-manager
